Jooq (#707)
* db-migration module

* merge main, postgres config

* add jooq generated classes to gitignore

* fix jooq configuration customizer

* migrate PhotoDao to PhotoAdapter

* remove generateSchemaSourceOnCompilation.set(false)

* refactor CountryDao to CountryAdapter

* configure gradle jvmargs

* migrate OAuth2AuthorizationDao to OAuth2AuthorizationAdapter

* WIP: migrate UserDao to UserAdapter

* finalize migration of UserDao to UserAdapter

* WIP: migrate InboxDao to InboxAdapter

* WIP: migrate InboxDao to InboxAdapter

* add InboxAdapterTest

* WIP: migrate StationDao to StationAdapter

* WIP: migrate StationDao to StationAdapter

* complete migration of StationDao to StationAdapter

+++ adapter/src/main/kotlin/org/railwaystations/rsapi/adapter/web/controller/StatisticController.kt
@@ -17,9 +17,9 @@ class StatisticController(private val getStatisticUseCase: GetStatisticUseCase) 
 }
 
 private fun Statistic.toDto() = StatisticDto(
-    total = total,
-    withPhoto = withPhoto,
-    withoutPhoto = withoutPhoto,
-    photographers = photographers,
+    total = total.toLong(),
+    withPhoto = withPhoto.toLong(),
+    withoutPhoto = withoutPhoto.toLong(),
+    photographers = photographers.toLong(),
     countryCode = countryCode
 )
